fix: scroll in app - add min-h-0 to flex chain (Layout, main)
Made-with: Cursor

+++ supabase/migrations/20260310_decrement_stock.sql
@@ -1,4 +1,4 @@
-CREATE OR REPLACE FUNCTION decrement_stock(p_item_id uuid, p_qty int)
+CREATE OR REPLACE FUNCTION decrement_stock(p_item_id uuid, p_qty numeric)
 RETURNS void AS $$
 BEGIN
   UPDATE inventory
